requirements.txt
# Core web framework
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx>=0.27.0

# AI & Orchestration
langchain-core>=0.2.0
langchain-openai>=0.1.3
langgraph>=0.1.0
chromadb>=0.4.24
langchain-chroma>=0.1.0
duckduckgo-search>=6.2.13

# MCP & Integrations
mcp>=1.0.0
